Consolidate agent modules into single enhanced agent implementation
